You Can Kill The Bats Now
This is some really sloppy code but it works.  I'll have to clean this up
a bit later.  I don't think it is entirely necessary to have an entire
group for the sword or to do the iterations that are being performed. I
just might keep it though if I decide I want to have more weapons.