Attacking Bats Now Does Damage Instead of Kill
Because there is no feedback for hitting a bat it seems a bit wonky but
you now only do damage to the bats instead of automatically killing them.
Next step will be adding some sort of feedback so that you can tell that
you actually hit one of the bats.